@@ -1,19 +1,31 @@
 {
   "name": "@auth0/nextjs-auth0",
-  "version": "1.9.2",
+  "version": "2.0.0",
   "description": "Next.js SDK for signing in with Auth0",
+  "exports": {
+    ".": "./dist/index.js",
+    "./client": "./dist/client/index.js",
+    "./edge": "./dist/edge.js",
+    "./testing": "./dist/helpers/testing.js"
+  },
   "main": "dist/index.js",
   "types": "dist/index.d.ts",
-  "browser": "dist/index.browser.js",
+  "browser": "dist/client/index.js",
   "directories": {
     "test": "tests"
   },
   "files": [
     "dist",
-    "src"
+    "src",
+    "client.js",
+    "client.d.ts",
+    "edge.js",
+    "edge.d.ts",
+    "testing.js",
+    "testing.d.ts"
   ],
   "engines": {
-    "node": "^10.13.0 || >=12.0.0"
+    "node": ">=12.19.0"
   },
   "scripts": {
     "clean": "rimraf dist",
@@ -85,20 +98,21 @@
     "cypress": "^11.2.0",
     "eslint": "^7.17.0",
     "eslint-config-prettier": "^8.1.0",
-    "eslint-import-resolver-typescript": "^2.3.0",
+    "eslint-import-resolver-typescript": "^3.5.2",
     "eslint-plugin-import": "^2.22.1",
     "eslint-plugin-jsx-a11y": "^6.4.1",
     "eslint-plugin-prettier": "^3.3.1",
     "eslint-plugin-react": "^7.22.0",
     "eslint-plugin-react-hooks": "^4.2.0",
     "jest": "^27.2.0",
-    "next": "^12.1.6",
+    "next": "^13.0.3",
     "nock": "^13.0.5",
     "oidc-provider": "^7.6.0",
+    "on-headers": "^1.0.2",
     "prettier": "^2.2.1",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
-    "react-test-renderer": "^17.0.1",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
+    "react-test-renderer": "^18.2.0",
     "rimraf": "^3.0.2",
     "start-server-and-test": "^1.11.7",
     "timekeeper": "^2.2.0",
