@@ -1,4 +1,3 @@
 module.exports = {
-  poweredByHeader: false,
-  target: 'serverless'
+  poweredByHeader: false
 };
