@@ -1,3 +1,8 @@
 import { handleAuth } from '@auth0/nextjs-auth0';
 
-export default handleAuth();
+export default handleAuth({
+  onError(req, res, error) {
+    console.error(error);
+    res.status(error.status || 500).end('Check the console for the error');
+  }
+});
