@@ -1,5 +1,6 @@
 import React from 'react';
-import { UserProfile, withPageAuthRequired } from '@auth0/nextjs-auth0';
+import { withPageAuthRequired } from '@auth0/nextjs-auth0';
+import { UserProfile } from '@auth0/nextjs-auth0/client';
 
 import Layout from '../components/layout';
 
@@ -9,11 +10,8 @@ export default function Profile({ user }: ProfileProps): React.ReactElement {
   return (
     <Layout>
       <h1>Profile</h1>
-
-      <div>
-        <h4>Profile (server rendered)</h4>
-        <pre data-testid="profile">{JSON.stringify(user, null, 2)}</pre>
-      </div>
+      <h4>Profile (server rendered)</h4>
+      <pre data-testid="profile">{JSON.stringify(user, null, 2)}</pre>
     </Layout>
   );
 }
