@@ -3,22 +3,25 @@ const next = require('next');
 const oidc = require('../../scripts/oidc-provider');
 
 const port = process.env.PORT || 3000;
-const app = next({ dev: true });
+const app = next({ dev: true, hostname: 'localhost', port });
 const handle = app.getRequestHandler();
 
-app.prepare().then(() => {
-  const server = express();
- 
-  server.use('/oidc', oidc({}).callback());
+app
+  .prepare()
+  .then(() => {
+    const server = express();
 
-  server.all('*', (req, res) => {
-    return handle(req, res);
-  });
+    server.use('/oidc', oidc({}).callback());
 
-  server.listen(port, (err) => {
-    if (err) throw err
-    console.log(`> Ready on http://localhost:${port}`);
+    server.all('*', (req, res) => {
+      return handle(req, res);
+    });
+
+    server.listen(port, (err) => {
+      if (err) throw err;
+      console.log(`> Ready on http://localhost:${port}`);
+    });
   })
-}).catch(err => {
+  .catch((err) => {
     console.log('Error:::::', err);
-});
+  });
