@@ -3,10 +3,10 @@ import urlJoin from 'url-join';
 import { strict as assert } from 'assert';
 import { Config, LoginOptions } from '../config';
 import TransientStore, { StoreOptions } from '../transient-store';
-import { encodeState } from '../hooks/get-login-state';
+import { encodeState } from '../utils/encoding';
 import { ClientFactory } from '../client';
 import createDebug from '../utils/debug';
-import { htmlSafe } from '../../utils/errors';
+import { htmlSafe } from '../utils/errors';
 
 const debug = createDebug('handlers');
 
@@ -50,23 +50,31 @@ export default function loginHandlerFactory(
     stateValue.nonce = transientHandler.generateNonce();
     stateValue.returnTo = stateValue.returnTo || opts.returnTo;
 
-    const usePKCE = (opts.authorizationParams.response_type as string).includes('code');
+    const responseType = opts.authorizationParams.response_type as string;
+    const usePKCE = responseType.includes('code');
     if (usePKCE) {
       debug('response_type includes code, the authorization request will use PKCE');
       stateValue.code_verifier = transientHandler.generateCodeVerifier();
     }
 
+    if (responseType !== config.authorizationParams.response_type) {
+      await transientHandler.save('response_type', req, res, {
+        ...transientOpts,
+        value: responseType
+      });
+    }
+
     const authParams = {
       ...opts.authorizationParams,
-      nonce: transientHandler.save('nonce', req, res, transientOpts),
-      state: transientHandler.save('state', req, res, {
+      nonce: await transientHandler.save('nonce', req, res, transientOpts),
+      state: await transientHandler.save('state', req, res, {
         ...transientOpts,
         value: encodeState(stateValue)
       }),
       ...(usePKCE
         ? {
             code_challenge: transientHandler.calculateCodeChallenge(
-              transientHandler.save('code_verifier', req, res, transientOpts)
+              await transientHandler.save('code_verifier', req, res, transientOpts)
             ),
             code_challenge_method: 'S256'
           }
@@ -81,7 +89,7 @@ export default function loginHandlerFactory(
     assert(/\bopenid\b/.test(authParams.scope as string), 'scope should contain "openid"');
 
     if (authParams.max_age) {
-      transientHandler.save('max_age', req, res, {
+      await transientHandler.save('max_age', req, res, {
         ...transientOpts,
         value: authParams.max_age.toString()
       });
