@@ -1,6 +1,7 @@
 import { IncomingMessage } from 'http';
-import { AuthorizationParameters as OidcAuthorizationParameters } from 'openid-client';
-import { LoginOptions, DeepPartial, getConfig as getBaseConfig } from './auth0-session';
+import type { AuthorizationParameters as OidcAuthorizationParameters } from 'openid-client';
+import type { LoginOptions } from './auth0-session/config';
+import { DeepPartial, get as getBaseConfig } from './auth0-session/get-config';
 
 /**
  * @category server
@@ -9,8 +10,9 @@ export interface BaseConfig {
   /**
    * The secret(s) used to derive an encryption key for the user identity in a session cookie and
    * to sign the transient cookies used by the login callback.
-   * Use a single string key or array of keys for an encrypted session cookie.
-   * You can also use the AUTH0_SECRET environment variable.
+   * Provide a single string secret, but if you want to rotate the secret you can provide an array putting
+   * the new secret first.
+   * You can also use the `AUTH0_SECRET` environment variable.
    */
   secret: string | Array<string>;
 
@@ -48,59 +50,59 @@ export interface BaseConfig {
    *   scope: 'openid profile email',
    *
    *   // Additional parameters
-   *   acr_value: "tenant:test-tenant",
-   *   custom_param: "custom-value"
+   *   acr_value: 'tenant:test-tenant',
+   *   custom_param: 'custom-value'
    * };
    * ```
    */
   authorizationParams: AuthorizationParameters;
 
   /**
-   * The root URL for the application router, eg https://localhost
-   * You can also use the AUTH0_BASE_URL environment variable.
-   * If you provide a domain, we will prefix it with `https://` - This can be useful when assigning it to
-   * `VERCEL_URL` for Vercel deploys
+   * The root URL for the application router, for example `https://localhost`.
+   * You can also use the `AUTH0_BASE_URL` environment variable.
+   * If you provide a domain, we will prefix it with `https://`. This can be useful when assigning it to
+   * `VERCEL_URL` for Vercel deploys.
    */
   baseURL: string;
 
   /**
    * The Client ID for your application.
-   * You can also use the AUTH0_CLIENT_ID environment variable.
+   * You can also use the `AUTH0_CLIENT_ID` environment variable.
    */
   clientID: string;
 
   /**
    * The Client Secret for your application.
    * Required when requesting access tokens.
-   * You can also use the AUTH0_CLIENT_SECRET environment variable.
+   * You can also use the `AUTH0_CLIENT_SECRET` environment variable.
    */
   clientSecret?: string;
 
   /**
    * Integer value for the system clock's tolerance (leeway) in seconds for ID token verification.`
-   * Default is 60
-   * You can also use the AUTH0_CLOCK_TOLERANCE environment variable.
+   * Defaults to `60` seconds.
+   * You can also use the `AUTH0_CLOCK_TOLERANCE` environment variable.
    */
   clockTolerance: number;
 
   /**
-   * Integer value for the http timeout in ms for authentication requests.
-   * Default is 5000
-   * You can also use the AUTH0_HTTP_TIMEOUT environment variable.
+   * Integer value for the HTTP timeout in milliseconds for authentication requests.
+   * Defaults to `5000` ms.
+   * You can also use the `AUTH0_HTTP_TIMEOUT` environment variable.
    */
   httpTimeout: number;
 
   /**
-   * To opt-out of sending the library and node version to your authorization server
-   * via the `Auth0-Client` header. Default is `true
-   * You can also use the AUTH0_ENABLE_TELEMETRY environment variable.
+   * Boolean value to opt-out of sending the library and node version to your authorization server
+   * via the `Auth0-Client` header. Defaults to `true`.
+   * You can also use the `AUTH0_ENABLE_TELEMETRY` environment variable.
    */
   enableTelemetry: boolean;
 
   /**
    * Function that returns an object with URL-safe state values for login.
    * Used for passing custom state parameters to your authorization server.
-   * Can also be passed in to {@link HandleLogin}
+   * Can also be passed in to {@link HandleLogin}.
    *
    * ```js
    * {
@@ -118,34 +120,34 @@ export interface BaseConfig {
 
   /**
    * Array value of claims to remove from the ID token before storing the cookie session.
-   * Default is `['aud', 'iss', 'iat', 'exp', 'nbf', 'nonce', 'azp', 'auth_time', 's_hash', 'at_hash', 'c_hash' ]`
-   * You can also use the AUTH0_IDENTITY_CLAIM_FILTER environment variable.
+   * Defaults to `['aud', 'iss', 'iat', 'exp', 'nbf', 'nonce', 'azp', 'auth_time', 's_hash', 'at_hash', 'c_hash']`.
+   * You can also use the `AUTH0_IDENTITY_CLAIM_FILTER` environment variable.
    */
   identityClaimFilter: string[];
 
   /**
-   * Boolean value to log the user out from the identity provider on application logout. Default is `true`
-   * You can also use the AUTH0_IDP_LOGOUT environment variable.
+   * Boolean value to log the user out from the identity provider on application logout. Defaults to `true`.
+   * You can also use the `AUTH0_IDP_LOGOUT` environment variable.
    */
   idpLogout: boolean;
 
   /**
-   * String value for the expected ID token algorithm. Default is 'RS256'
-   * You can also use the AUTH0_ID_TOKEN_SIGNING_ALG environment variable.
+   * String value for the expected ID token algorithm. Defaults to 'RS256'.
+   * You can also use the `AUTH0_ID_TOKEN_SIGNING_ALG` environment variable.
    */
   idTokenSigningAlg: string;
 
   /**
-   * REQUIRED. The root URL for the token issuer with no trailing slash.
-   * This is `https://` plus your Auth0 domain
-   * You can also use the AUTH0_ISSUER_BASE_URL environment variable.
+   * **REQUIRED** The root URL for the token issuer with no trailing slash.
+   * This is `https://` plus your Auth0 domain.
+   * You can also use the `AUTH0_ISSUER_BASE_URL` environment variable.
    */
   issuerBaseURL: string;
 
   /**
    * Set a fallback cookie with no `SameSite` attribute when `response_mode` is `form_post`.
    * The default `response_mode` for this SDK is `query` so this defaults to `false`
-   * You can also use the AUTH0_LEGACY_SAME_SITE_COOKIE environment variable.
+   * You can also use the `AUTH0_LEGACY_SAME_SITE_COOKIE` environment variable.
    */
   legacySameSiteCookie: boolean;
 
@@ -157,14 +159,14 @@ export interface BaseConfig {
      * Either a relative path to the application or a valid URI to an external domain.
      * This value must be registered on the authorization server.
      * The user will be redirected to this after a logout has been performed.
-     * You can also use the AUTH0_POST_LOGOUT_REDIRECT environment variable.
+     * You can also use the `AUTH0_POST_LOGOUT_REDIRECT` environment variable.
      */
     postLogoutRedirect: string;
 
     /**
      * Relative path to the application callback to process the response from the authorization server.
-     * Defaults to `/api/auth/callback`
-     * You can also use the AUTH0_CALLBACK environment variable.
+     * Defaults to `/api/auth/callback`.
+     * You can also use the `AUTH0_CALLBACK` environment variable.
      */
     callback: string;
   };
@@ -179,26 +181,26 @@ export interface SessionConfig {
   /**
    * String value for the cookie name used for the internal session.
    * This value must only include letters, numbers, and underscores.
-   * Default is `appSession`.
-   * You can also use the AUTH0_SESSION_NAME environment variable.
+   * Defaults to `appSession`.
+   * You can also use the `AUTH0_SESSION_NAME` environment variable.
    */
   name: string;
 
   /**
-   * If you want your session duration to be rolling, eg reset everytime the
-   * user is active on your site, set this to a `true`. If you want the session
-   * duration to be absolute, where the user is logged out a fixed time after login,
-   * regardless of activity, set this to `false`
-   * Default is `true`.
-   * You can also use the AUTH0_SESSION_ROLLING environment variable.
+   * If you want your session duration to be rolling, resetting everytime the
+   * user is active on your site, set this to `true`. If you want the session
+   * duration to be absolute, where the user gets logged out a fixed time after login
+   * regardless of activity, set this to `false`.
+   * Defaults to `true`.
+   * You can also use the `AUTH0_SESSION_ROLLING` environment variable.
    */
   rolling: boolean;
 
   /**
    * Integer value, in seconds, for application session rolling duration.
    * The amount of time for which the user must be idle for then to be logged out.
-   * Should be false when rolling is false.
-   * Default is 86400 seconds (1 day).
+   * Should be `false` when rolling is `false`.
+   * Defaults to `86400` seconds (1 day).
    * You can also use the AUTH0_SESSION_ROLLING_DURATION environment variable.
    */
   rollingDuration: number | false;
@@ -207,10 +209,17 @@ export interface SessionConfig {
    * Integer value, in seconds, for application absolute rolling duration.
    * The amount of time after the user has logged in that they will be logged out.
    * Set this to `false` if you don't want an absolute duration on your session.
-   * Default is 604800 seconds (7 days).
-   * You can also use the AUTH0_SESSION_ABSOLUTE_DURATION environment variable.
+   * Defaults to `604800` seconds (7 days).
+   * You can also use the `AUTH0_SESSION_ABSOLUTE_DURATION` environment variable.
    */
   absoluteDuration: boolean | number;
+
+  /**
+   * Boolean value to store the ID token in the session. Storing it can make the session cookie too
+   * large.
+   * Defaults to `true`.
+   */
+  storeIDToken: boolean;
 
   cookie: CookieConfig;
 }
@@ -223,43 +232,43 @@ export interface SessionConfig {
 export interface CookieConfig {
   /**
    * Domain name for the cookie.
-   * You can also use the AUTH0_COOKIE_DOMAIN environment variable.
+   * You can also use the `AUTH0_COOKIE_DOMAIN` environment variable.
    */
   domain?: string;
 
   /**
    * Path for the cookie.
-   * This defaults to `/`
+   * Defaults to `/`.
    * You should change this to be more restrictive if you application shares a domain with other apps.
-   * You can also use the AUTH0_COOKIE_PATH environment variable.
+   * You can also use the `AUTH0_COOKIE_PATH` environment variable.
    */
   path?: string;
 
   /**
-   * Set to true to use a transient cookie (cookie without an explicit expiration).
-   * Default is `false`
-   * You can also use the AUTH0_COOKIE_TRANSIENT environment variable.
+   * Set to `true` to use a transient cookie (cookie without an explicit expiration).
+   * Defaults to `false`.
+   * You can also use the `AUTH0_COOKIE_TRANSIENT` environment variable.
    */
   transient: boolean;
 
   /**
    * Flags the cookie to be accessible only by the web server.
    * Defaults to `true`.
-   * You can also use the AUTH0_COOKIE_HTTP_ONLY environment variable.
+   * You can also use the `AUTH0_COOKIE_HTTP_ONLY` environment variable.
    */
   httpOnly: boolean;
 
   /**
    * Marks the cookie to be used over secure channels only.
    * Defaults to the protocol of {@link BaseConfig.baseURL}.
-   * You can also use the AUTH0_COOKIE_SECURE environment variable.
+   * You can also use the `AUTH0_COOKIE_SECURE` environment variable.
    */
   secure?: boolean;
 
   /**
-   * Value of the SameSite Set-Cookie attribute.
-   * Defaults to "lax" but will be adjusted based on {@link AuthorizationParameters.response_type}.
-   * You can also use the AUTH0_COOKIE_SAME_SITE environment variable.
+   * Value of the SameSite `Set-Cookie` attribute.
+   * Defaults to `lax` but will be adjusted based on {@link AuthorizationParameters.response_type}.
+   * You can also use the `AUTH0_COOKIE_SAME_SITE` environment variable.
    */
   sameSite: 'lax' | 'strict' | 'none';
 }
@@ -272,7 +281,13 @@ export interface CookieConfig {
  * @category Server
  */
 export interface AuthorizationParameters extends OidcAuthorizationParameters {
+  /**
+   * A space-separated list of scopes that will be requested during authentication. For example,
+   * `openid profile email offline_access`.
+   * Defaults to `openid profile email`.
+   */
   scope: string;
+
   response_mode: 'query' | 'form_post';
   response_type: 'id_token' | 'code id_token' | 'code';
 }
@@ -285,13 +300,15 @@ export interface NextConfig extends Pick<BaseConfig, 'identityClaimFilter'> {
    * Log users in to a specific organization.
    *
    * This will specify an `organization` parameter in your user's login request and will add a step to validate
-   * the `org_id` claim in your user's ID Token.
+   * the `org_id` claim in your user's ID token.
    *
-   * If your app supports multiple organizations, you should take a look at {@Link AuthorizationParams.organization}
+   * If your app supports multiple organizations, you should take a look at {@link AuthorizationParams.organization}.
    */
   organization?: string;
   routes: {
+    callback: string;
     login: string;
+    unauthorized: string;
   };
 }
 
@@ -300,11 +317,11 @@ export interface NextConfig extends Pick<BaseConfig, 'identityClaimFilter'> {
  *
  * The Server part of the SDK can be configured in 2 ways.
  *
- * ### 1. Environmental Variables
+ * ### 1. Environment Variables
  *
  * The simplest way to use the SDK is to use the named exports ({@link HandleAuth}, {@link HandleLogin},
  * {@link HandleLogout}, {@link HandleCallback}, {@link HandleProfile}, {@link GetSession}, {@link GetAccessToken},
- * {@link WithApiAuthRequired} and {@link WithPageAuthRequired}), eg:
+ * {@link WithApiAuthRequired}, and {@link WithPageAuthRequired}).
  *
  * ```js
  * // pages/api/auth/[...auth0].js
@@ -314,41 +331,41 @@ export interface NextConfig extends Pick<BaseConfig, 'identityClaimFilter'> {
  * ```
  *
  * When you use these named exports, an instance of the SDK is created for you which you can configure using
- * environmental variables:
+ * environment variables:
  *
  * ### Required
  *
- * - `AUTH0_SECRET`: See {@link secret}
- * - `AUTH0_ISSUER_BASE_URL`: See {@link issuerBaseURL}
- * - `AUTH0_BASE_URL`: See {@link baseURL}
- * - `AUTH0_CLIENT_ID`: See {@link clientID}
- * - `AUTH0_CLIENT_SECRET`: See {@link clientSecret}
+ * - `AUTH0_SECRET`: See {@link secret}.
+ * - `AUTH0_ISSUER_BASE_URL`: See {@link issuerBaseURL}.
+ * - `AUTH0_BASE_URL`: See {@link baseURL}.
+ * - `AUTH0_CLIENT_ID`: See {@link clientID}.
+ * - `AUTH0_CLIENT_SECRET`: See {@link clientSecret}.
  *
  * ### Optional
  *
- * - `AUTH0_CLOCK_TOLERANCE`: See {@link clockTolerance}
- * - `AUTH0_HTTP_TIMEOUT`: See {@link httpTimeout}
- * - `AUTH0_ENABLE_TELEMETRY`: See {@link enableTelemetry}
- * - `AUTH0_IDP_LOGOUT`: See {@link idpLogout}
- * - `AUTH0_ID_TOKEN_SIGNING_ALG`: See {@link idTokenSigningAlg}
- * - `AUTH0_LEGACY_SAME_SITE_COOKIE`: See {@link legacySameSiteCookie}
- * - `AUTH0_IDENTITY_CLAIM_FILTER`: See {@link identityClaimFilter}
- * - `NEXT_PUBLIC_AUTH0_LOGIN`: See {@link NextConfig.routes}
- * - `AUTH0_CALLBACK`: See {@link BaseConfig.routes}
- * - `AUTH0_POST_LOGOUT_REDIRECT`: See {@link BaseConfig.routes}
- * - `AUTH0_AUDIENCE`: See {@link BaseConfig.authorizationParams}
- * - `AUTH0_SCOPE`: See {@link BaseConfig.authorizationParams}
- * - `AUTH0_ORGANIZATION`: See {@link NextConfig.organization}
- * - `AUTH0_SESSION_NAME`: See {@link SessionConfig.name}
- * - `AUTH0_SESSION_ROLLING`: See {@link SessionConfig.rolling}
- * - `AUTH0_SESSION_ROLLING_DURATION`: See {@link SessionConfig.rollingDuration}
- * - `AUTH0_SESSION_ABSOLUTE_DURATION`: See {@link SessionConfig.absoluteDuration}
- * - `AUTH0_COOKIE_DOMAIN`: See {@link CookieConfig.domain}
- * - `AUTH0_COOKIE_PATH`: See {@link CookieConfig.path}
- * - `AUTH0_COOKIE_TRANSIENT`: See {@link CookieConfig.transient}
- * - `AUTH0_COOKIE_HTTP_ONLY`: See {@link CookieConfig.httpOnly}
- * - `AUTH0_COOKIE_SECURE`: See {@link CookieConfig.secure}
- * - `AUTH0_COOKIE_SAME_SITE`: See {@link CookieConfig.sameSite}
+ * - `AUTH0_CLOCK_TOLERANCE`: See {@link clockTolerance}.
+ * - `AUTH0_HTTP_TIMEOUT`: See {@link httpTimeout}.
+ * - `AUTH0_ENABLE_TELEMETRY`: See {@link enableTelemetry}.
+ * - `AUTH0_IDP_LOGOUT`: See {@link idpLogout}.
+ * - `AUTH0_ID_TOKEN_SIGNING_ALG`: See {@link idTokenSigningAlg}.
+ * - `AUTH0_LEGACY_SAME_SITE_COOKIE`: See {@link legacySameSiteCookie}.
+ * - `AUTH0_IDENTITY_CLAIM_FILTER`: See {@link identityClaimFilter}.
+ * - `NEXT_PUBLIC_AUTH0_LOGIN`: See {@link NextConfig.routes}.
+ * - `AUTH0_CALLBACK`: See {@link BaseConfig.routes}.
+ * - `AUTH0_POST_LOGOUT_REDIRECT`: See {@link BaseConfig.routes}.
+ * - `AUTH0_AUDIENCE`: See {@link BaseConfig.authorizationParams}.
+ * - `AUTH0_SCOPE`: See {@link BaseConfig.authorizationParams}.
+ * - `AUTH0_ORGANIZATION`: See {@link NextConfig.organization}.
+ * - `AUTH0_SESSION_NAME`: See {@link SessionConfig.name}.
+ * - `AUTH0_SESSION_ROLLING`: See {@link SessionConfig.rolling}.
+ * - `AUTH0_SESSION_ROLLING_DURATION`: See {@link SessionConfig.rollingDuration}.
+ * - `AUTH0_SESSION_ABSOLUTE_DURATION`: See {@link SessionConfig.absoluteDuration}.
+ * - `AUTH0_COOKIE_DOMAIN`: See {@link CookieConfig.domain}.
+ * - `AUTH0_COOKIE_PATH`: See {@link CookieConfig.path}.
+ * - `AUTH0_COOKIE_TRANSIENT`: See {@link CookieConfig.transient}.
+ * - `AUTH0_COOKIE_HTTP_ONLY`: See {@link CookieConfig.httpOnly}.
+ * - `AUTH0_COOKIE_SECURE`: See {@link CookieConfig.secure}.
+ * - `AUTH0_COOKIE_SAME_SITE`: See {@link CookieConfig.sameSite}.
  *
  * ### 2. Create your own instance using {@link InitAuth0}
  *
@@ -373,7 +390,7 @@ export interface NextConfig extends Pick<BaseConfig, 'identityClaimFilter'> {
  * return auth0.handleAuth();
  * ```
  *
- * **Note** If you use {@link InitAuth0}, you should *not* use the other named exports as they will use a different
+ * **IMPORTANT** If you use {@link InitAuth0}, you should *not* use the other named exports as they will use a different
  * instance of the SDK. Also note - this is for the server side part of the SDK - you will always use named exports for
  * the front end components: {@Link UserProvider}, {@Link UseUser} and the
  * front end version of {@Link WithPageAuthRequired}
@@ -440,6 +457,7 @@ export const getConfig = (params: ConfigParameters = {}): { baseConfig: BaseConf
   const AUTH0_SESSION_ROLLING = process.env.AUTH0_SESSION_ROLLING;
   const AUTH0_SESSION_ROLLING_DURATION = process.env.AUTH0_SESSION_ROLLING_DURATION;
   const AUTH0_SESSION_ABSOLUTE_DURATION = process.env.AUTH0_SESSION_ABSOLUTE_DURATION;
+  const AUTH0_SESSION_STORE_ID_TOKEN = process.env.AUTH0_SESSION_STORE_ID_TOKEN;
   const AUTH0_COOKIE_DOMAIN = process.env.AUTH0_COOKIE_DOMAIN;
   const AUTH0_COOKIE_PATH = process.env.AUTH0_COOKIE_PATH;
   const AUTH0_COOKIE_TRANSIENT = process.env.AUTH0_COOKIE_TRANSIENT;
@@ -484,6 +502,7 @@ export const getConfig = (params: ConfigParameters = {}): { baseConfig: BaseConf
         AUTH0_SESSION_ABSOLUTE_DURATION && isNaN(Number(AUTH0_SESSION_ABSOLUTE_DURATION))
           ? bool(AUTH0_SESSION_ABSOLUTE_DURATION)
           : num(AUTH0_SESSION_ABSOLUTE_DURATION),
+      storeIDToken: bool(AUTH0_SESSION_STORE_ID_TOKEN),
       ...baseParams.session,
       cookie: {
         domain: AUTH0_COOKIE_DOMAIN,
@@ -504,7 +523,8 @@ export const getConfig = (params: ConfigParameters = {}): { baseConfig: BaseConf
   const nextConfig = {
     routes: {
       ...baseConfig.routes,
-      login: baseParams.routes?.login || getLoginUrl()
+      login: baseParams.routes?.login || getLoginUrl(),
+      unauthorized: baseParams.routes?.unauthorized || '/api/auth/401'
     },
     identityClaimFilter: baseConfig.identityClaimFilter,
     organization: organization || AUTH0_ORGANIZATION
