@@ -1,12 +1,13 @@
+import { NextApiHandler, NextApiRequest, NextApiResponse } from 'next';
 import { HandleLogin } from './login';
 import { HandleLogout } from './logout';
 import { HandleCallback } from './callback';
 import { HandleProfile } from './profile';
-import { NextApiHandler, NextApiRequest, NextApiResponse } from 'next';
+import { HandlerError } from '../utils/errors';
 
 /**
  * If you want to add some custom behavior to the default auth handlers, you can pass in custom handlers for
- * `login`, `logout`, `callback` and `profile` eg
+ * `login`, `logout`, `callback`, and `profile`. For example:
  *
  * ```js
  * // pages/api/auth/[...auth0].js
@@ -23,25 +24,55 @@ import { NextApiHandler, NextApiRequest, NextApiResponse } from 'next';
  *     } catch (error) {
  *       // Add you own custom error logging.
  *       errorReporter(error);
- *       res.status(error.status || 500).end(error.message);
+ *       res.status(error.status || 500).end();
  *     }
  *   }
  * });
  * ```
  *
+ * Alternatively, you can customize the default handlers without overriding them. For example:
+ *
+ * ```js
+ * // pages/api/auth/[...auth0].js
+ * import { handleAuth, handleLogin } from '@auth0/nextjs-auth0';
+ *
+ * export default handleAuth({
+ *   login: handleLogin({
+ *     authorizationParams: { customParam: 'foo' } // Pass in custom params
+ *   })
+ * });
+ * ```
+ *
+ * You can also create new handlers by customizing the default ones. For example:
+ *
+ * ```js
+ * // pages/api/auth/[...auth0].js
+ * import { handleAuth, handleLogin } from '@auth0/nextjs-auth0';
+ *
+ * export default handleAuth({
+ *   signup: handleLogin({
+ *     authorizationParams: { screen_hint: 'signup' }
+ *   })
+ * });
+ * ```
+ *
  * @category Server
  */
-export interface Handlers {
-  login: HandleLogin;
-  logout: HandleLogout;
-  callback: HandleCallback;
-  profile: HandleProfile;
-}
+export type Handlers = ApiHandlers | ErrorHandlers;
+
+type ApiHandlers = {
+  [key: string]: NextApiHandler;
+};
+
+type ErrorHandlers = {
+  onError?: OnError;
+};
 
 /**
  * The main way to use the server SDK.
  *
- * Simply set the environment variables per {@link Config} then create the file `pages/api/auth/[...auth0].js`, eg
+ * Simply set the environment variables per {@link ConfigParameters} then create the file
+ * `pages/api/auth/[...auth0].js`. For example:
  *
  * ```js
  * // pages/api/auth/[...auth0].js
@@ -50,30 +81,60 @@ export interface Handlers {
  * export default handleAuth();
  * ```
  *
- * This will create 4 handlers for the following urls:
+ * This will create 5 handlers for the following urls:
  *
- * - `/api/auth/login`: log the user in to your app by redirecting them to your Identity Provider.
- * - `/api/auth/callback`: The page that your Identity Provider will redirect the user back to on login.
+ * - `/api/auth/login`: log the user in to your app by redirecting them to your identity provider.
+ * - `/api/auth/callback`: The page that your identity provider will redirect the user back to on login.
  * - `/api/auth/logout`: log the user out of your app.
- * - `/api/auth/me`: View the user profile JSON (used by the {@link UseUser} hook)
+ * - `/api/auth/me`: View the user profile JSON (used by the {@link UseUser} hook).
+ * - `/api/auth/unauthorized`: Returns a 401 for use by {@link WithMiddlewareAuthRequired} when protecting API routes.
  *
  * @category Server
  */
-export type HandleAuth = (userHandlers?: Partial<Handlers>) => NextApiHandler;
+export type HandleAuth = (userHandlers?: Handlers) => NextApiHandler;
+
+/**
+ * Error handler for the default auth routes.
+ *
+ * Use this to define an error handler for all the default routes in a single place. For example:
+ *
+ * ```js
+ * export default handleAuth({
+ *   onError(req, res, error) {
+ *     errorLogger(error);
+ *     // You can finish the response yourself if you want to customize
+ *     // the status code or redirect the user
+ *     // res.writeHead(302, {
+ *     //     Location: '/custom-error-page'
+ *     // });
+ *     // res.end();
+ *   }
+ * });
+ * ```
+ *
+ * @category Server
+ */
+export type OnError = (req: NextApiRequest, res: NextApiResponse, error: HandlerError) => Promise<void> | void;
 
 /**
  * @ignore
  */
-const wrapErrorHandling = (fn: NextApiHandler): NextApiHandler => async (
-  req: NextApiRequest,
-  res: NextApiResponse
-): Promise<void> => {
-  try {
-    await fn(req, res);
-  } catch (error) {
-    console.error(error);
-    res.status(error.status || 500).end(error.message);
-  }
+const defaultOnError: OnError = (_req, res, error) => {
+  console.error(error);
+  res.status(error.status || 500).end();
+};
+
+/**
+ * This is a handler for use by {@link WithMiddlewareAuthRequired} when protecting an API route.
+ * Middleware can't return a response body, so an unauthorized request for an API route
+ * needs to rewrite to this handler.
+ * @ignore
+ */
+const unauthorized: NextApiHandler = (_req, res) => {
+  res.status(401).json({
+    error: 'not_authenticated',
+    description: 'The user does not have an active session or is not authenticated'
+  });
 };
 
 /**
@@ -90,32 +151,35 @@ export default function handlerFactory({
   handleCallback: HandleCallback;
   handleProfile: HandleProfile;
 }): HandleAuth {
-  return (userHandlers: Partial<Handlers> = {}): NextApiHandler => {
-    const { login, logout, callback, profile } = {
-      login: wrapErrorHandling(handleLogin),
-      logout: wrapErrorHandling(handleLogout),
-      callback: wrapErrorHandling(handleCallback),
-      profile: wrapErrorHandling(handleProfile),
-      ...userHandlers
+  return ({ onError, ...handlers }: Handlers = {}): NextApiHandler<void> => {
+    const customHandlers: ApiHandlers = {
+      login: handleLogin,
+      logout: handleLogout,
+      callback: handleCallback,
+      me: (handlers as ApiHandlers).profile || handleProfile,
+      401: unauthorized,
+      ...handlers
     };
     return async (req, res): Promise<void> => {
       let {
         query: { auth0: route }
       } = req;
 
-      route = Array.isArray(route) ? route[0] : route;
+      route = Array.isArray(route) ? route[0] : /* c8 ignore next */ route;
 
-      switch (route) {
-        case 'login':
-          return login(req, res) as void;
-        case 'logout':
-          return logout(req, res) as void;
-        case 'callback':
-          return callback(req, res) as void;
-        case 'me':
-          return profile(req, res) as void;
-        default:
+      try {
+        const handler = route && customHandlers.hasOwnProperty(route) && customHandlers[route];
+        if (handler) {
+          await handler(req, res);
+        } else {
           res.status(404).end();
+        }
+      } catch (error) {
+        await (onError || defaultOnError)(req, res, error as HandlerError);
+        if (!res.writableEnded) {
+          // 200 is the default, so we assume it has not been set in the custom error handler if it equals 200
+          res.status(res.statusCode === 200 ? 500 : res.statusCode).end();
+        }
       }
     };
   };
