@@ -2,4 +2,4 @@ export { default as callbackHandler, HandleCallback, CallbackOptions, AfterCallb
 export { default as loginHandler, HandleLogin, LoginOptions, GetLoginState } from './login';
 export { default as logoutHandler, HandleLogout, LogoutOptions } from './logout';
 export { default as profileHandler, HandleProfile, ProfileOptions, AfterRefetch } from './profile';
-export { default as handlerFactory, Handlers, HandleAuth } from './auth';
+export { default as handlerFactory, Handlers, HandleAuth, OnError } from './auth';
