@@ -1,66 +1,68 @@
 import { IncomingMessage, ServerResponse } from 'http';
 import { NextApiRequest, NextApiResponse } from 'next';
-import { TokenSet } from 'openid-client';
-import onHeaders from 'on-headers';
+import type { TokenSet } from 'openid-client';
 import { Config, SessionCache as ISessionCache, CookieStore } from '../auth0-session';
 import Session, { fromJson, fromTokenSet } from './session';
 
-type NextApiOrPageRequest = IncomingMessage | NextApiRequest;
-type NextApiOrPageResponse = ServerResponse | NextApiResponse;
+export default class SessionCache<
+  Req extends object = IncomingMessage | NextApiRequest, // eslint-disable-line @typescript-eslint/ban-types
+  Res extends object = ServerResponse | NextApiResponse // eslint-disable-line @typescript-eslint/ban-types
+> implements ISessionCache<Req, Res, Session>
+{
+  private cache: WeakMap<Req, Session | null>;
+  private iatCache: WeakMap<Req, number | undefined>;
 
-export default class SessionCache implements ISessionCache {
-  private cache: WeakMap<NextApiOrPageRequest, Session | null>;
-  private iatCache: WeakMap<NextApiOrPageRequest, number | undefined>;
-
-  constructor(private config: Config, private cookieStore: CookieStore) {
+  constructor(private config: Config, private cookieStore: CookieStore<Req, Res>) {
     this.cache = new WeakMap();
     this.iatCache = new WeakMap();
   }
 
-  init(req: NextApiOrPageRequest, res: NextApiOrPageResponse, autoSave = true): void {
+  private async init(req: Req, res: Res, autoSave = true): Promise<void> {
     if (!this.cache.has(req)) {
-      const [json, iat] = this.cookieStore.read(req);
+      const [json, iat] = await this.cookieStore.read(req);
       this.iatCache.set(req, iat);
       this.cache.set(req, fromJson(json));
-      if (autoSave) {
-        onHeaders(res, () => this.save(req, res));
+      if (this.config.session.rolling && autoSave) {
+        await this.save(req, res);
       }
     }
   }
 
-  save(req: NextApiOrPageRequest, res: NextApiOrPageResponse): void {
-    this.cookieStore.save(req, res, this.cache.get(req), this.iatCache.get(req));
+  async save(req: Req, res: Res): Promise<void> {
+    await this.cookieStore.save(req, res, this.cache.get(req), this.iatCache.get(req));
   }
 
-  create(req: NextApiOrPageRequest, res: NextApiOrPageResponse, session: Session): void {
+  async create(req: Req, res: Res, session: Session): Promise<void> {
     this.cache.set(req, session);
-    onHeaders(res, () => this.save(req, res));
+    await this.save(req, res);
   }
 
-  delete(req: NextApiOrPageRequest, res: NextApiOrPageResponse): void {
-    this.init(req, res);
+  async delete(req: Req, res: Res): Promise<void> {
+    await this.init(req, res, false);
     this.cache.set(req, null);
+    await this.save(req, res);
   }
 
-  isAuthenticated(req: NextApiOrPageRequest, res: NextApiOrPageResponse): boolean {
-    this.init(req, res);
+  async isAuthenticated(req: Req, res: Res): Promise<boolean> {
+    await this.init(req, res);
     const session = this.cache.get(req);
     return !!session?.user;
   }
 
-  getIdToken(req: NextApiOrPageRequest, res: NextApiOrPageResponse): string | undefined {
-    this.init(req, res);
+  async getIdToken(req: Req, res: Res): Promise<string | undefined> {
+    await this.init(req, res);
     const session = this.cache.get(req);
     return session?.idToken;
   }
 
-  set(req: NextApiOrPageRequest, res: NextApiOrPageResponse, session: Session | null): void {
-    this.init(req, res);
+  async set(req: Req, res: Res, session: Session | null): Promise<void> {
+    await this.init(req, res, false);
     this.cache.set(req, session);
+    await this.save(req, res);
   }
 
-  get(req: NextApiOrPageRequest, res: NextApiOrPageResponse): Session | null | undefined {
-    this.init(req, res);
+  async get(req: Req, res: Res): Promise<Session | null | undefined> {
+    await this.init(req, res);
     return this.cache.get(req);
   }
 
