@@ -1,6 +1,7 @@
 import { IncomingMessage, ServerResponse } from 'http';
 import { NextApiRequest, NextApiResponse } from 'next';
 import { SessionCache, Session } from '../session';
+import { assertReqRes } from '../utils/assert';
 
 /**
  * Get the user's session from the request.
@@ -10,13 +11,14 @@ import { SessionCache, Session } from '../session';
 export type GetSession = (
   req: IncomingMessage | NextApiRequest,
   res: ServerResponse | NextApiResponse
-) => Session | null | undefined;
+) => Promise<Session | null | undefined>;
 
 /**
  * @ignore
  */
 export default function sessionFactory(sessionCache: SessionCache): GetSession {
-  return (req, res): Session | null | undefined => {
+  return (req, res) => {
+    assertReqRes(req, res);
     return sessionCache.get(req, res);
   };
 }
