@@ -1,1 +1,1 @@
-export default '1.9.2';
+export default '2.0.0';
