@@ -1,6 +1,6 @@
 import { AddressInfo } from 'net';
 import { createServer, get as getRequest, IncomingMessage, ServerResponse } from 'http';
-import { getAll, get, set, clear } from '../../../src/auth0-session/utils/cookies';
+import NodeCookies from '../../../src/auth0-session/utils/cookies';
 
 const setup = (): Promise<[IncomingMessage, ServerResponse, Function]> =>
   new Promise((resolve) => {
@@ -25,27 +25,31 @@ describe('cookie', () => {
   it('should get all cookies', async () => {
     const [req, , teardown] = await setup();
     req.headers.cookie = 'foo=bar; bar=baz;';
-    expect(getAll(req)).toMatchObject({ foo: 'bar', bar: 'baz' });
+    expect(new NodeCookies().getAll(req)).toMatchObject({ foo: 'bar', bar: 'baz' });
     await teardown();
   });
 
   it('should get a cookie by name', async () => {
     const [req, , teardown] = await setup();
     req.headers.cookie = 'foo=bar; bar=baz;';
-    expect(get(req, 'foo')).toEqual('bar');
+    expect(new NodeCookies().getAll(req)['foo']).toEqual('bar');
     await teardown();
   });
 
   it('should set a cookie', async () => {
     const [, res, teardown] = await setup();
-    set(res, 'foo', 'bar');
+    const setter = new NodeCookies();
+    setter.set('foo', 'bar');
+    setter.commit(res);
     expect(res.getHeader('Set-Cookie')).toEqual(['foo=bar']);
     await teardown();
   });
 
   it('should set a cookie with opts', async () => {
     const [, res, teardown] = await setup();
-    set(res, 'foo', 'bar', { httpOnly: true, sameSite: 'strict' });
+    const setter = new NodeCookies();
+    setter.set('foo', 'bar', { httpOnly: true, sameSite: 'strict' });
+    setter.commit(res);
     expect(res.getHeader('Set-Cookie')).toEqual(['foo=bar; HttpOnly; SameSite=Strict']);
     await teardown();
   });
@@ -53,22 +57,38 @@ describe('cookie', () => {
   it('should not overwrite existing set cookie', async () => {
     const [, res, teardown] = await setup();
     res.setHeader('Set-Cookie', 'foo=bar');
-    set(res, 'baz', 'qux');
+    const setter = new NodeCookies();
+    setter.set('baz', 'qux');
+    setter.commit(res);
     expect(res.getHeader('Set-Cookie')).toEqual(['foo=bar', 'baz=qux']);
     await teardown();
   });
 
-  it('should not overwrite existing set cookie array', async () => {
+  it('should override existing cookies that equal name', async () => {
     const [, res, teardown] = await setup();
-    set(res, 'foo', 'bar');
-    set(res, 'baz', 'qux');
-    expect(res.getHeader('Set-Cookie')).toEqual(['foo=bar', 'baz=qux']);
+    res.setHeader('Set-Cookie', ['foo=bar', 'baz=qux']);
+    const setter = new NodeCookies();
+    setter.set('foo', 'qux');
+    setter.commit(res, 'foo');
+    expect(res.getHeader('Set-Cookie')).toEqual(['baz=qux', 'foo=qux']);
+    await teardown();
+  });
+
+  it('should override existing cookies that match name', async () => {
+    const [, res, teardown] = await setup();
+    res.setHeader('Set-Cookie', ['foo.1=bar', 'foo.2=baz']);
+    const setter = new NodeCookies();
+    setter.set('foo', 'qux');
+    setter.commit(res, 'foo');
+    expect(res.getHeader('Set-Cookie')).toEqual(['foo=qux']);
     await teardown();
   });
 
   it('should clear cookies', async () => {
     const [, res, teardown] = await setup();
-    clear(res, 'foo');
+    const setter = new NodeCookies();
+    setter.clear('foo');
+    setter.commit(res);
     expect(res.getHeader('Set-Cookie')).toEqual(['foo=; Max-Age=0']);
     await teardown();
   });
