@@ -48,6 +48,7 @@ describe('config params', () => {
         rolling: true,
         rollingDuration: 86400,
         absoluteDuration: 604800,
+        storeIDToken: true,
         cookie: {
           domain: undefined,
           path: '/',
@@ -91,7 +92,8 @@ describe('config params', () => {
       routes: {
         login: '/api/auth/login',
         callback: '/api/auth/callback',
-        postLogoutRedirect: ''
+        postLogoutRedirect: '',
+        unauthorized: '/api/auth/401'
       },
       organization: undefined
     });
@@ -107,7 +109,8 @@ describe('config params', () => {
         AUTH0_COOKIE_HTTP_ONLY: 'on',
         AUTH0_COOKIE_SAME_SITE: 'lax',
         AUTH0_COOKIE_SECURE: 'ok',
-        AUTH0_SESSION_ABSOLUTE_DURATION: 'no'
+        AUTH0_SESSION_ABSOLUTE_DURATION: 'no',
+        AUTH0_SESSION_STORE_ID_TOKEN: '0'
       }).baseConfig
     ).toMatchObject({
       auth0Logout: false,
@@ -116,6 +119,7 @@ describe('config params', () => {
       legacySameSiteCookie: false,
       session: {
         absoluteDuration: false,
+        storeIDToken: false,
         cookie: {
           httpOnly: true,
           sameSite: 'lax',
@@ -182,6 +186,7 @@ describe('config params', () => {
         },
         session: {
           absoluteDuration: 100,
+          storeIDToken: false,
           cookie: {
             transient: false
           },
@@ -201,6 +206,7 @@ describe('config params', () => {
       },
       session: {
         absoluteDuration: 100,
+        storeIDToken: false,
         cookie: {
           transient: false
         },
