@@ -1,1 +1,2 @@
-export default (global as any).handleAuth();
+// eslint-disable-next-line @typescript-eslint/explicit-module-boundary-types
+export default (...args: any) => global.handleAuth?.()(...args);
