@@ -1,17 +1,8 @@
 import { parse } from 'cookie';
-import { parse as parseUrl, URL } from 'url';
+import { parse as parseUrl } from 'url';
 import { withoutApi } from '../fixtures/default-settings';
+import { get } from '../auth0-session/fixtures/helpers';
 import { setup, teardown, login } from '../fixtures/setup';
-import { IncomingMessage } from 'http';
-
-jest.mock('../../src/utils/assert', () => ({
-  assertReqRes(req: IncomingMessage) {
-    if (req.url?.includes('error=')) {
-      const url = new URL(req.url, 'http://example.com');
-      throw new Error(url.searchParams.get('error') as string);
-    }
-  }
-}));
 
 describe('logout handler', () => {
   afterEach(teardown);
@@ -20,20 +11,44 @@ describe('logout handler', () => {
     const baseUrl = await setup(withoutApi);
     const cookieJar = await login(baseUrl);
 
-    const { status, headers } = await fetch(`${baseUrl}/api/auth/logout`, {
-      redirect: 'manual',
-      headers: {
-        cookie: cookieJar.getCookieStringSync(baseUrl)
-      }
+    const {
+      res: { statusCode, headers }
+    } = await get(baseUrl, '/api/auth/logout', {
+      cookieJar,
+      fullResponse: true
     });
 
-    expect(status).toBe(302);
-    expect(parseUrl(headers.get('location') as string, true)).toMatchObject({
+    expect(statusCode).toBe(302);
+    expect(parseUrl(headers['location'], true)).toMatchObject({
       protocol: 'https:',
       host: 'acme.auth0.local',
       query: {
         returnTo: 'http://www.acme.com',
         client_id: '__test_client_id__'
+      },
+      pathname: '/v2/logout'
+    });
+  });
+
+  test('should pass logout params to the identity provider', async () => {
+    const baseUrl = await setup(withoutApi, { logoutOptions: { logoutParams: { foo: 'bar' } } });
+    const cookieJar = await login(baseUrl);
+
+    const {
+      res: { statusCode, headers }
+    } = await get(baseUrl, '/api/auth/logout', {
+      cookieJar,
+      fullResponse: true
+    });
+
+    expect(statusCode).toBe(302);
+    expect(parseUrl(headers['location'], true)).toMatchObject({
+      protocol: 'https:',
+      host: 'acme.auth0.local',
+      query: {
+        returnTo: 'http://www.acme.com',
+        client_id: '__test_client_id__',
+        foo: 'bar'
       },
       pathname: '/v2/logout'
     });
@@ -46,15 +61,15 @@ describe('logout handler', () => {
     });
     const cookieJar = await login(baseUrl);
 
-    const { status, headers } = await fetch(`${baseUrl}/api/auth/logout`, {
-      redirect: 'manual',
-      headers: {
-        cookie: cookieJar.getCookieStringSync(baseUrl)
-      }
+    const {
+      res: { statusCode, headers }
+    } = await get(baseUrl, '/api/auth/logout', {
+      cookieJar,
+      fullResponse: true
     });
 
-    expect(status).toBe(302);
-    expect(parseUrl(headers.get('location') as string, true).query).toMatchObject({
+    expect(statusCode).toBe(302);
+    expect(parseUrl(headers['location'], true).query).toMatchObject({
       returnTo: 'https://www.foo.bar'
     });
   });
@@ -65,15 +80,15 @@ describe('logout handler', () => {
     });
     const cookieJar = await login(baseUrl);
 
-    const { status, headers } = await fetch(`${baseUrl}/api/auth/logout`, {
-      redirect: 'manual',
-      headers: {
-        cookie: cookieJar.getCookieStringSync(baseUrl)
-      }
+    const {
+      res: { statusCode, headers }
+    } = await get(baseUrl, '/api/auth/logout', {
+      cookieJar,
+      fullResponse: true
     });
 
-    expect(status).toBe(302);
-    expect(parseUrl(headers.get('location') as string)).toMatchObject({
+    expect(statusCode).toBe(302);
+    expect(parseUrl(headers['location'])).toMatchObject({
       host: 'my-end-session-endpoint',
       pathname: '/logout'
     });
@@ -85,25 +100,17 @@ describe('logout handler', () => {
     });
     const cookieJar = await login(baseUrl);
 
-    const res = await fetch(`${baseUrl}/api/auth/logout`, {
-      redirect: 'manual',
-      headers: {
-        cookie: cookieJar.getCookieStringSync(baseUrl)
-      }
+    const {
+      res: { headers }
+    } = await get(baseUrl, '/api/auth/logout', {
+      cookieJar,
+      fullResponse: true
     });
 
-    expect(parse(res.headers.get('set-cookie') as string)).toMatchObject({
+    expect(parse(headers['set-cookie'][0])).toMatchObject({
       appSession: '',
       'Max-Age': '0',
       Path: '/'
     });
-  });
-
-  test('should escape html in errors', async () => {
-    const baseUrl = await setup(withoutApi);
-
-    const res = await fetch(`${baseUrl}/api/auth/logout?error=%3Cscript%3Ealert(%27xss%27)%3C%2Fscript%3E`);
-
-    expect(await res.text()).toEqual('&lt;script&gt;alert(&#39;xss&#39;)&lt;/script&gt;');
   });
 });
