@@ -4,10 +4,8 @@ module.exports = {
   exclude: [
     './src/auth0-session/**',
     './src/session/cache.ts',
-    './src/frontend/use-config.tsx',
-    './src/utils/!(errors.ts)',
-    './src/index.ts',
-    './src/index.browser.ts'
+    './src/client/use-config.tsx',
+    './src/utils/!(errors.ts)'
   ],
   excludeExternals: true,
   excludePrivate: true,
